flaskapi/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
pre-commit==3.7.1
pytest==7.0.1
google-cloud-aiplatform==1.35.0
pandas==2.0.3
langchain==0.0.315
chromadb==0.4.14
gcsfs
jupyterlab